first commit of aws module utils including AnsibleAWSModule and moving fail_json_aws there
