+++ test/units/module_utils/ec2/test_aws.py
@@ -103,86 +103,3 @@ class RetryTestCase(unittest.TestCase):
 
         self.assertEqual(self.counter, 1)
 
-
-class ErrorReportingTestcase(unittest.TestCase):
-
-    def test_botocore_exception_reports_nicely_via_fail_json_aws(self):
-
-        basic._ANSIBLE_ARGS=to_bytes(json.dumps({'ANSIBLE_MODULE_ARGS': {}}))
-        module = AnsibleModule(argument_spec = dict(
-            fail_mode = dict(type='list', default=['success'])
-        ))
-
-        fail_json_double=Mock()
-        err_msg = {'Error': {'Code': 'FakeClass.FakeError'}}
-        with patch.object(basic.AnsibleModule, 'fail_json', fail_json_double):
-            try:
-                raise botocore.exceptions.ClientError(err_msg, 'Could not find you')
-            except Exception as e:
-                print("exception is " + str(e))
-                module.fail_json_aws(e, msg="Fake failure for testing boto exception messages")
-
-        assert(len(fail_json_double.mock_calls) > 0), "failed to call fail_json when should have"
-        assert(len(fail_json_double.mock_calls) < 2), "called fail_json multiple times when once would do"
-        assert("test_botocore_exception_reports_nicely"
-               in fail_json_double.mock_calls[0][2]["exception"]), \
-               "exception traceback doesn't include correct function, fail call was actually: " \
-                + str(fail_json_double.mock_calls[0])
-
-        assert("Fake failure for testing boto exception messages:" in fail_json_double.mock_calls[0][2]["msg"]), \
-            "error message doesn't include the local message; was: " \
-            + str(fail_json_double.mock_calls[0])
-        assert("Could not find you" in fail_json_double.mock_calls[0][2]["msg"]), \
-            "error message doesn't include the botocore exception message; was: " \
-            + str(fail_json_double.mock_calls[0])
-        try:
-            fail_json_double.mock_calls[0][2]["error"]
-        except KeyError:
-            raise Exception("error was missing; call was: " + str(fail_json_double.mock_calls[0]))
-        assert("FakeClass.FakeError" == fail_json_double.mock_calls[0][2]["error"]["code"]), \
-            "Failed to find error/code; was: " + str(fail_json_double.mock_calls[0])
-
-
-
-    def test_botocore_exception_without_response_reports_nicely_via_fail_json_aws(self):
-
-        basic._ANSIBLE_ARGS=to_bytes(json.dumps({'ANSIBLE_MODULE_ARGS': {}}))
-        module = AnsibleModule(argument_spec = dict(
-            fail_mode = dict(type='list', default=['success'])
-        ))
-
-        fail_json_double=Mock()
-        err_msg = None
-        with patch.object(basic.AnsibleModule, 'fail_json', fail_json_double):
-            try:
-                raise botocore.exceptions.ClientError(err_msg, 'Could not find you')
-            except Exception as e:
-                print("exception is " + str(e))
-                module.fail_json_aws(e, msg="Fake failure for testing boto exception messages again")
-
-        assert(len(fail_json_double.mock_calls) > 0), "failed to call fail_json when should have"
-        assert(len(fail_json_double.mock_calls) < 2), \
-            "called fail_json multiple times when once would do"
-
-        assert("test_botocore_exception_without_response_reports_nicely_via_fail_json_aws"
-               in fail_json_double.mock_calls[0][2]["exception"]), \
-               "exception traceback doesn't include correct function, fail call was actually: " \
-               + str(fail_json_double.mock_calls[0])
-
-        assert("Fake failure for testing boto exception messages again:" in fail_json_double.mock_calls[0][2]["msg"]), \
-            "error message doesn't include the local message; was: " \
-            + str(fail_json_double.mock_calls[0])
-
-        # I would have thought this should work, however the botocore exception comes back with
-        # "argument of type 'NoneType' is not iterable" so it's probably not really designed
-        # to handle "None" as an error response.
-        #
-        # assert("Could not find you" in fail_json_double.mock_calls[0][2]["msg"]), \
-        #    "error message doesn't include the botocore exception message; was: " \
-        #    + str(fail_json_double.mock_calls[0])
-
-
-# TODO:
-#  - an exception without a message
-#  - plain boto exception
-#  - socket errors and other standard things.
